Rename files to align with rules_python_gazelle
