+++ gazelle/js/gazelle_test.go
@@ -17,7 +17,7 @@ limitations under the License.
 // https://github.com/bazelbuild/bazel-skylib/blob/f80bc733d4b9f83d427ce3442be2e07427b2cc8d/gazelle/bzl/BUILD.
 // It was modified for the needs of this extension.
 
-package js_test
+package js
 
 import (
 	"bytes"
